Add a means of looking up all settings for a specified key
